move framework commit message generation inside gitCommit
